call_multiple_callback_functions() fix.
The issue was the following:

 It seems that your intention was the following:

If the num_parallel parameter is not specified, then execute the functions one by one, starting the next function when the previous one is ended.

If the num_parallel parameter is specified, then allows to execute up to num_parallel functions at the same time.

In both cases the "main" callback function should be called one time only, when all the functions are completed (i.e. their callbacks are called).

Presently the "main" callback is called num_parallel times. It seems like an error.